Made it so new flashcards could be added to sets + minor file reorganisation
