+++ WebAplicationTestMVCSOL/WebAplicationTestMVC/Views/Home/AddFlashcard.cshtml
@@ -1,13 +1,11 @@
-﻿@page
-
-@using WebAplicationTestMVC.Pages;
-@model AddFlashcardModel
+﻿@model WebAplicationTestMVC.Models.StudySet;
 @{
   
 }
 
 <h1>Make a Flashcard</h1>
-<form method="post">
+@using (Html.BeginForm("SubmitNewFlashcard", "Home", FormMethod.Post))
+{
     <label for="question">FlashCard Question:</label><br>
     <textarea id="question" name="question" rows="7" cols="100" placeholder="Where does England keep their armies?" required></textarea>
     <br>
@@ -16,7 +14,7 @@
     <textarea id="answer" name="answer" rows="7" cols="100" placeholder="In their sleevies!" required></textarea>
     <br>
     <br>
-    <button type="submit" class="btn btn-primary" name="submitButton" value="submit" asp-page-handler="SubmitNewFlashcard">Submit</button>
-    <a class="btn btn-primary" asp-area="" asp-controller="Home" asp-action="StudySet">Back</a>
+    <button type="submit" value="@Model.studySetName" name="studySetName" class="btn btn-primary">Submit</button>
+    <a class="btn btn-primary" asp-area="" asp-controller="Home" asp-action="StudySets" asp-route-studySetName="@Model.studySetName">Back</a>
     
-</form>
+}
